requirements.txt
torch
transformers
pydantic
accelerate
sentencepiece
python-dotenv
protobuf
pika